requirements.txt
asgiref==3.2.10
catkin==0.7.23
Django==3.1.2
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
fuzzywuzzy==0.18.0
PyJWT==1.7.1
python-Levenshtein==0.12.0
pytz==2020.1
sqlparse==0.4.1
